EfficientNetB7 implementation
Created a python script that creates the EfficientNetB7 model with some additional FC layers in order to use transfer learning. There is also a notebook file for training the model. I also had to do some changes on the resnet model since it wasn't pushed in the last issue, and instead got deleted.